Refactor Find components to shared folder
