--- app/components/find/utility/advice_component/view.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-module Find
-  module Utility
-    module AdviceComponent
-      class View < ViewComponent::Base
-        include ::ViewHelper
-
-        attr_reader :title
-
-        def initialize(title:)
-          super
-          @title = title
-        end
-      end
-    end
-  end
-end
