+++ app/components/shared/advice_component/view.html.erb
