+++ app/components/shared/advice_component/view.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Shared
+  module AdviceComponent
+    class View < ViewComponent::Base
+      include ::ViewHelper
+
+      attr_reader :title
+
+      def initialize(title:)
+        super
+        @title = title
+      end
+    end
+  end
+end
