+++ app/components/shared/courses/financial_support/bursary_component/view.html.erb
