+++ app/components/shared/courses/financial_support/fees_and_financial_support_component/view.html.erb
@@ -5,17 +5,17 @@
     <%= render Find::Courses::FeesComponent::View.new(course) %>
   <% end %>
 
-  <%= render Find::Utility::AdviceComponent::View.new(title: "Financial support from the government") do %>
+  <%= render Shared::AdviceComponent::View.new(title: "Financial support from the government") do %>
     <% if salaried? %>
-      <%= render partial: "find/courses/financial_support/salaried" %>
+      <%= render partial: "shared/courses/financial_support/salaried" %>
     <% elsif excluded_from_bursary? %>
-      <%= render partial: "find/courses/financial_support/loan" %>
+      <%= render partial: "shared/courses/financial_support/loan" %>
     <% elsif bursary_only? %>
-      <%= render Find::Courses::FinancialSupport::BursaryComponent::View.new(course) %>
+      <%= render Shared::Courses::FinancialSupport::BursaryComponent::View.new(course) %>
     <% elsif has_scholarship_and_bursary? %>
-      <%= render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course) %>
+      <%= render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course) %>
     <% else %>
-      <%= render partial: "find/courses/financial_support/loan" %>
+      <%= render partial: "shared/courses/financial_support/loan" %>
     <% end %>
   <% end %>
