+++ app/components/shared/courses/financial_support/scholarship_and_bursary_component/view.html.erb
