+++ app/views/find/courses/show.html.erb
@@ -46,7 +46,7 @@
       <%= render partial: "find/courses/salary", locals: { course: @course } %>
     <% end %>
 
-    <%= render Find::Courses::FinancialSupport::FeesAndFinancialSupportComponent::View.new(@course) %>
+    <%= render Shared::Courses::FinancialSupport::FeesAndFinancialSupportComponent::View.new(@course) %>
 
     <% if @provider.train_with_us.present? || @course.about_accrediting_provider.present? %>
       <%= render partial: "find/courses/about_the_provider", locals: { course: @course } %>
