+++ app/views/publish/courses/fees_and_financial_support/_guidance.html.erb
@@ -6,17 +6,17 @@
     </span>
   </summary>
   <div class="govuk-!-margin-top-4" />
-  <%= render Find::Utility::AdviceComponent::View.new(title: "Financial support from the government") do %>
+  <%= render Shared::AdviceComponent::View.new(title: "Financial support from the government") do %>
     <% if @course.salaried? %>
-      <%= render partial: "find/courses/financial_support/salaried" %>
+      <%= render partial: "shared/courses/financial_support/salaried" %>
     <% elsif @course.excluded_from_bursary? %>
-      <%= render partial: "find/courses/financial_support/loan" %>
+      <%= render partial: "shared/courses/financial_support/loan" %>
     <% elsif @course.bursary_only? %>
-      <%= render Find::Courses::FinancialSupport::BursaryComponent::View.new(@course) %>
+      <%= render Shared::Courses::FinancialSupport::BursaryComponent::View.new(@course) %>
     <% elsif @course.has_scholarship_and_bursary? %>
-      <%= render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(@course) %>
+      <%= render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(@course) %>
     <% else %>
-      <%= render partial: "find/courses/financial_support/loan" %>
+      <%= render partial: "shared/courses/financial_support/loan" %>
     <% end %>
   <% end %>
 </details>
