+++ app/views/publish/courses/school_placements/_guidance.html.erb
@@ -8,7 +8,7 @@
       </span>
     </summary>
     <div class="govuk-!-margin-top-4" />
-    <%= render Find::Utility::AdviceComponent::View.new(title: t("publish.providers.school_placements.edit.where_you_will_train")) do %>
+    <%= render Shared::AdviceComponent::View.new(title: t("publish.providers.school_placements.edit.where_you_will_train")) do %>
       <%= t("publish.providers.school_placements.edit.university_guidance_html") %>
     <% end %>
   </details>
@@ -20,7 +20,7 @@
       </span>
     </summary>
     <div class="govuk-!-margin-top-4" />
-    <%= render Find::Utility::AdviceComponent::View.new(title: t("publish.providers.school_placements.edit.where_you_will_train")) do %>
+    <%= render Shared::AdviceComponent::View.new(title: t("publish.providers.school_placements.edit.where_you_will_train")) do %>
       <%= t("publish.providers.school_placements.edit.scitt_guidance_html") %>
     <% end %>
   </details>
