+++ spec/components/shared/courses/financial_support/bursary_component/view_preview.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module Find
+module Shared
   module Courses
     module FinancialSupport
       module BursaryComponent
@@ -11,7 +11,7 @@ module Find
                                 program_type: 'higher_education_programme',
                                 level: 'further_education',
                                 subjects: [Subject.new(id: 49, type: 'SecondarySubject', subject_code: 'C7', subject_name: 'Physical education with an EBacc subject', financial_incentive: FinancialIncentive.new(bursary_amount: 3000))])
-            render Find::Courses::FinancialSupport::BursaryComponent::View.new(course)
+            render Shared::Courses::FinancialSupport::BursaryComponent::View.new(course)
           end
         end
       end
