+++ spec/components/shared/courses/financial_support/fees_and_financial_support_component/view_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe Find::Courses::FinancialSupport::FeesAndFinancialSupportComponent::View, type: :component do
+describe Shared::Courses::FinancialSupport::FeesAndFinancialSupportComponent::View, type: :component do
   context 'Salaried courses' do
     it 'renders salaried course section if the course has a salary' do
       course = build(:course, funding_type: 'salary').decorate
