+++ spec/components/shared/courses/financial_support/scholarship_and_bursary_component/view_preview.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module Find
+module Shared
   module Courses
     module FinancialSupport
       module ScholarshipAndBursaryComponent
@@ -10,7 +10,7 @@ module Find
                                                                                                                                                                                   bursary_amount: 3000,
                                                                                                                                                                                   early_career_payments: 2000))])
 
-            render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
+            render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
           end
 
           def chemistry_with_scholarship_early_career_and_scholarship_body
@@ -18,7 +18,7 @@ module Find
                                                                                                                                                                                     bursary_amount: 3000,
                                                                                                                                                                                     early_career_payments: 2000))])
 
-            render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
+            render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
           end
 
           def computing_with_scholarship_early_career_and_scholarship_body
@@ -26,7 +26,7 @@ module Find
                                                                                                                                                                                     bursary_amount: 3000,
                                                                                                                                                                                     early_career_payments: 2000))])
 
-            render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
+            render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
           end
 
           def maths_with_scholarship_early_career_and_scholarship_body
@@ -34,7 +34,7 @@ module Find
                                                                                                                                                                                 bursary_amount: 3000,
                                                                                                                                                                                 early_career_payments: 2000))])
 
-            render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
+            render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
           end
 
           def french_with_scholarship_early_career_and_scholarship_body
@@ -42,7 +42,7 @@ module Find
                                                                                                                                                                                  bursary_amount: 3000,
                                                                                                                                                                                  early_career_payments: 2000))])
 
-            render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
+            render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
           end
 
           def german_with_scholarship_early_career_and_scholarship_body
@@ -50,7 +50,7 @@ module Find
                                                                                                                                                                                  bursary_amount: 3000,
                                                                                                                                                                                  early_career_payments: 2000))])
 
-            render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
+            render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
           end
 
           def spanish_with_scholarship_early_career_and_scholarship_body
@@ -58,21 +58,21 @@ module Find
                                                                                                                                                                                   bursary_amount: 3000,
                                                                                                                                                                                   early_career_payments: 2000))])
 
-            render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
+            render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
           end
 
           def without_scholarship_and_scholarship_body
             course = Course.new(subjects: [Subject.new(id: 22, type: 'SecondarySubject', subject_code: 'F8', subject_name: 'Physics', financial_incentive: FinancialIncentive.new(scholarship: 2000,
                                                                                                                                                                                   bursary_amount: 3000))])
 
-            render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
+            render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
           end
 
           def with_scholarship_only
             course = Course.new(subjects: [Subject.new(id: 22, type: 'SecondarySubject', subject_code: 'F8', subject_name: 'Art', financial_incentive: FinancialIncentive.new(scholarship: 2000,
                                                                                                                                                                               bursary_amount: 3000))])
 
-            render Find::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
+            render Shared::Courses::FinancialSupport::ScholarshipAndBursaryComponent::View.new(course.decorate)
           end
         end
       end
